requirements.txt
transformers==4.48.3
llama-index==0.12.16
llama-index-llms-huggingface==0.4.2
llama-index-embeddings-huggingface==0.5.1
llama-index-llms-huggingface-api==0.3.1
huggingface-hub==0.28.1
fastapi==0.115.8
uvicorn==0.34.0
streamlit==1.42.1
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic-settings==2.7.1
pydantic[email]
SQLAlchemy==2.0.38
pytest==8.3.4
python-jose==3.3.0
passlib==1.7.4
